UART library module
The UART can be used in 'cooked' (the default) or 'raw' mode.
The problem with 'cooked' mode is that it waits forever until a line has
been entered. This is useful if you're waiting for user input, but if
you're receiving data from a device, you don't want that.
The problem with 'raw' mode however is that it gives you exactly what
you ask of it - a buffer that may or may not be filled. And there is no
end-of-line parsing performed.
So this module sets the UART to 'raw' mode - but then provides a
ReadLine function as well.