NMEA-0183 protocol decoder module
The NMEA-0183 protocol is used by GPS devices to report what it decodes
from the GPS satellites.
The protocol is in ASCII, and usually received from a serial line of
some kind.
This module decodes the received packets into data structures for
programmatic processing elsewhere.
The only GPS module that I have access to is SparkFun's GP-20U7
56-channel chip/antenna combination - just perfect for an Intel® Edison!